Clean up and rename service
